refactor: correct categories
