+++ crates/store-dir/src/lib.rs
@@ -1,9 +1,9 @@
-mod index;
+mod cas_file;
+mod index_file;
 mod prune;
 mod store_dir;
-mod write_file;
 
-pub use index::*;
+pub use cas_file::*;
+pub use index_file::*;
 pub use prune::*;
 pub use store_dir::*;
-pub use write_file::*;
